refactor: create competition participations module
