+++ src/services/api/modules.ts
@@ -1,1 +1,6 @@
-export type TModuleName = 'players' | 'clubs' | 'teams' | 'competition-groups'
+export type TModuleName =
+  | 'players'
+  | 'clubs'
+  | 'teams'
+  | 'competition-groups'
+  | 'competition-participations'
